Refactoring.
A number of functions have been moved from the cmd package to the gpbckpconfig package.

For some functions, the function type has been changed from a local type to a public one. They can be used in other libraries.

The division by types for auxiliary functions has been performed: working with a file, working with a database, general functions.