Remove show-all flag for backup-info command.
Changed the behavior for the backup-info command.
Now, by default, active backups or backups with the deletion status "In progress" are displayed.
When additional flags are specified, the necessary backups are added to the displayed list: deleted (--show-deleted) or failed (--show-failed). To display all backups with any status, the options should be used together.

This way of displaying is much more convenient than the previous one.