Add e2e tets for report-info command.
This test compares the contents of the report file from the src_data directory with the result of the report-info command.

Also some refactoring compose file and Makefile. Now, when running e2e tests, a complete test environment is always prepared for each team. So that there are no intersections.