Add report-info for local backups for history yaml db.
Added basic functional for report-info command for local history db. In this commit - only for yaml format.
The new --backup-dir flag was added.

For connection to local GP cluster  and execution queries lib/pq and jmoiron/sqlx are used.

For local backups the following logic are applied:
  * If the --backup-dir option is specified, the report will be searched in provided path.
  * If the --backup-dir option is not specified, the utility try to connect to local cluster and get master data directory.
    If this information is available, the report will be in master data directory.
  * If backup is not local, the error will be returned.

For non local backups the following logic are applied:
  * If the --plugin-config option is specified, the report will be searched in provided location.
  * If backup is local, the error will be returned.