Refactor backup delete process.
gpbackup  save information in $MASTER_DATA_DIRECTORY/backups. When deleting a backup using a plugin, it is necessary to clear local backup information on the master.

Added deletion of files on the master when deleting a backup using a plugin. Also, performed a small refactoring with the addition of handlers for error handling.

There may be cases when a backup hangs in the "In Progress" status. Additionally, backups in the "Failure" status save information about themselves. You can get a report for them, and they also need to be deleted.

The conditions in the checkBackupCanBeUsed() function have been changed. The conditions in checkBackupCanBeUsed() function have been changed to work with backups that have not been deleted. This allows for the deletion of backups in "Failure" or "In Progress" statuses.